backend/pyproject.toml
[tool.poetry]
name = "news-app"
version = "0.1.0"
description = ""
authors = ["Simon Foucek <sfoucek@redhat.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.115.12"}
sqlmodel = "^0.0.24"
psycopg = {extras = ["binary", "pool"], version = "^3.2.6"}
alembic = "^1.15.2"
slowapi = "^0.1.9"
sqlalchemy = "^2.0.40"
instructor = "^1.7.9"
google-generativeai = "^0.8.5"
jsonref = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
